make the contents of locales/en/common.json and locales/ja/common.json the same
