+++ web/locales/en/common.json
@@ -1,5 +1,5 @@
 {
-  "follow_us_on_twitter": "follow us on twitter",
+  "follow_us_on_twitter": "公式twitterアカウント( @mercaridevjp )にて情報を発信していきます。ぜひフォローをお願いします。",
   "about": {
     "p1": "Mercari Tech Conf（MTC）は、株式会社メルカリをはじめとするメルカリグループ各社が、これから目指す方向や、これから取り組む技術的なチャレンジについてご紹介するエンジニア向けの技術カンファレンスです。",
     "p2": "今年で第2回となる「Mercari Tech Conf 2018」では、テーマを「Evolution（変化）」として、この1年間でメルカリグループ内に起こった、もしくは起こりつつある変化をご紹介します。",
