+++ web/utils/index.tsx
@@ -12,7 +12,7 @@ export const omitText = (text: string, maxLength: number = 20): string => {
  * 渡されたものを指定された数だけ
  * 複製して配列で返却
  */
-export const duplicate = <T>(n: number) => {
+export function duplicate<T>(n: number) {
   return (item: T): T[] => {
     const list: T[] = [];
     for (let i = 0; i < n; i++) {
@@ -20,7 +20,7 @@ export const duplicate = <T>(n: number) => {
     }
     return list;
   };
-};
+}
 
 /**
  * 渡された回数だけ関数を実行する
@@ -32,3 +32,17 @@ export const times = (n: number) => {
     }
   };
 };
+
+/**
+ * string[]を<br />でつなげる
+ */
+export function joinWithBr(texts: string[]): React.ReactNode[] {
+  const result: React.ReactNode[] = [];
+  for (let i = 0; i < texts.length; i++) {
+    result.push(texts[i]);
+    result.push(<br key={i} />);
+  }
+
+  result.splice(result.length - 1, 1);
+  return result;
+}
